+++ BMTextRenderer.h
@@ -12,14 +12,14 @@ class BMTextRenderer : public UIBase
 { 
 public:
 	BMTextRenderer(TEXTURE_ID textureID, std::string text,
-					Pivot pivot, int x, int y);
+					Pivot pivot, float x, float y);
 	void Render() override;
 
 	inline void SetSpacing(float spacing) { m_spacing = spacing; };
-	inline void SetSize(int size) { m_size = size; };
+	inline void SetSize(float size) { m_size = size; };
 	inline void SetText(std::string newText) { m_text = newText; };
 
-	inline int GetSpacing(int ascii) 
+	inline float GetSpacing(int ascii) 
 	{ 
 		return ascii == 32 ? m_spacing/2 : m_spacing;
 	};
@@ -27,8 +27,8 @@ public:
 private:
 
 	void ComputeTextSize();
-	int GetOffSetX();
-	int GetOffSetY();
+	float GetOffSetX();
+	float GetOffSetY();
 
 	const int m_startAsciiValue = 32;
 	const int m_maxCol = 26;
@@ -42,8 +42,8 @@ private:
 	Vector2 m_pos;
 	Pivot m_pivot;
 
-	int m_textSize = 0;
-	int m_size = 16;
+	float m_textSize = 0;
+	float m_size = 16;
 	float m_spacing = 16;
 };
